refactor: Error 관련 코드 global 하위 패키지로 이동
